Test: Small refac and add fingerprint
