+++ Adjust/test-app-core/build.gradle
@@ -19,7 +19,7 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(dir: 'libs', include: ['*.jar', '*.aar'])
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation "androidx.constraintlayout:constraintlayout:2.1.1"
     implementation 'com.google.android.gms:play-services-ads-identifier:17.1.0'
